Campaign ui uses GUIRadioButtonGroup

+++ Barotrauma/BarotraumaClient/Source/Items/Components/Machines/Steering.cs
@@ -142,9 +142,9 @@ namespace Barotrauma.Items.Components
             };
 
             GUIRadioButtonGroup modes = new GUIRadioButtonGroup();
-            modes.AddRadioButton(Mode.AutoPilot, autopilotTickBox);
-            modes.AddRadioButton(Mode.Manual, manualTickBox);
-            modes.Selected = Mode.Manual;
+            modes.AddRadioButton((int)Mode.AutoPilot, autopilotTickBox);
+            modes.AddRadioButton((int)Mode.Manual, manualTickBox);
+            modes.Selected = (int)Mode.Manual;
             
             var autoPilotControls = new GUIFrame(new RectTransform(new Vector2(1.0f, 0.6f), paddedControlContainer.RectTransform), "InnerFrame");
             var paddedAutoPilotControls = new GUILayoutGroup(new RectTransform(new Vector2(0.8f), autoPilotControls.RectTransform, Anchor.Center))
@@ -246,11 +246,11 @@ namespace Barotrauma.Items.Components
             autoPilotControlsDisabler = new GUIFrame(new RectTransform(Vector2.One, autoPilotControls.RectTransform), "InnerFrame");
 
             GUIRadioButtonGroup destinations = new GUIRadioButtonGroup();
-            destinations.AddRadioButton(Destination.MaintainPos, maintainPosTickBox);
-            destinations.AddRadioButton(Destination.LevelStart, levelStartTickBox);
-            destinations.AddRadioButton(Destination.LevelEnd, levelEndTickBox);
-            destinations.Selected = maintainPos        ? Destination.MaintainPos :
-                                    levelStartSelected ? Destination.LevelStart  : Destination.LevelEnd;
+            destinations.AddRadioButton((int)Destination.MaintainPos, maintainPosTickBox);
+            destinations.AddRadioButton((int)Destination.LevelStart, levelStartTickBox);
+            destinations.AddRadioButton((int)Destination.LevelEnd, levelEndTickBox);
+            destinations.Selected = (int)(maintainPos        ? Destination.MaintainPos :
+                                          levelStartSelected ? Destination.LevelStart  : Destination.LevelEnd);
             
             string steeringVelX = TextManager.Get("SteeringVelocityX");
             string steeringVelY = TextManager.Get("SteeringVelocityY");
